Add sorting and grouping props to Timeline component
The `orderBy` prop allows the user to specify whether to order the data
in ascending or descending order. Defaults to ascending order.

The `groupBy` prop allows the user to specify whether to group the data
by days, weeks, or months. Defaults to days.

To make this easier for the user objects are provided that contain these
parameters:
  * `groupByOptions` for grouping options.
  * `orderByOptions` for ordering options.

This PR also removes some of the redundant props passed to the Timeline
group and item components.